feat: entites 파일 분리 (Layers - Slices - Segments) 1. entities/post/model - post model 분리
